Refactor group CRUD tests into per-command modules
Rather than a singular test module for all of the various group
operations, a module per operation covers these better.

In the course of the refactor, a session-scoped autouse fixture was
moved into the module which uses it. (Rather than being in a sibling
module.)
