+++ tests/functional/groups/test_group_member_commands.py
@@ -13,6 +13,132 @@ from globus_sdk._testing import (
 
 
 @pytest.fixture(autouse=True, scope="session")
+def _register_get_group_responses():
+    group_id = "efdab3ca-cff1-11e4-9b86-123139260d4e"
+
+    register_response_set(
+        "get_group?include=memberships",
+        dict(
+            default=dict(
+                service="groups",
+                path=f"/groups/{group_id}",
+                json={
+                    "description": "Ipso facto",
+                    "enforce_session": False,
+                    "group_type": "regular",
+                    "id": group_id,
+                    "memberships": [
+                        {
+                            "group_id": group_id,
+                            "identity_id": "ae332d86-d274-11e5-b885-b31714a110e9",
+                            "membership_fields": {
+                                "department": "Globus Testing",
+                                "email": "sirosen@globus.org",
+                                "field_of_science": "CS",
+                                "institution": "Computation Institute",
+                                "phone": "867-5309",
+                            },
+                            "role": "admin",
+                            "status": "active",
+                            "username": "sirosen@globusid.org",
+                        },
+                        {
+                            "group_id": group_id,
+                            "identity_id": "508e5ef6-cb9b-11e5-abe1-431ce3f42be1",
+                            "membership_fields": {},
+                            "role": "member",
+                            "status": "invited",
+                            "username": "sirosen@xsede.org",
+                        },
+                        {
+                            "group_id": group_id,
+                            "identity_id": "ae2f7f60-d274-11e5-b879-afc598dd59d4",
+                            "membership_fields": {
+                                "institution": "University of Chicago",
+                                "name": "Bryce Allen",
+                                "department": "Globus",
+                            },
+                            "role": "member",
+                            "status": "active",
+                            "username": "ballen@globusid.org",
+                        },
+                        {
+                            "group_id": group_id,
+                            "identity_id": "b0e8f24a-d274-11e5-8c98-8fd1e61c0a76",
+                            "membership_fields": {
+                                "current_project_name": "Petrel support",
+                                "department": "UChicago",
+                            },
+                            "role": "member",
+                            "status": "rejected",
+                            "username": "smartin@globusid.org",
+                        },
+                        {
+                            "group_id": group_id,
+                            "identity_id": "6b487878-d2a1-11e5-b689-a7dd99513a65",
+                            "membership_fields": {
+                                "department": (
+                                    "Columbia University department "
+                                    "of Witchcraft and History"
+                                ),
+                            },
+                            "role": "member",
+                            "status": "active",
+                            "username": "jss2253@columbia.edu",
+                        },
+                        {
+                            "group_id": group_id,
+                            "identity_id": "ae2a1750-d274-11e5-b867-e74762c29f57",
+                            "membership_fields": {},
+                            "role": "member",
+                            "status": "invited",
+                            "username": "bjmc@globusid.org",
+                        },
+                    ],
+                    "name": "Claptrap Presents Claptrap's Rough Riders",
+                    "parent_id": None,
+                    "policies": {
+                        "authentication_assurance_timeout": 28800,
+                        "group_members_visibility": "managers",
+                        "group_visibility": "private",
+                        "is_high_assurance": False,
+                        "join_requests": False,
+                        "signup_fields": [],
+                    },
+                    "session_limit": 28800,
+                    "session_timeouts": {
+                        "ae341a98-d274-11e5-b888-dbae3a8ba545": {
+                            "expire_time": "2022-02-08T06:05:54+00:00",
+                            "expires_in": 0,
+                        }
+                    },
+                },
+                metadata={
+                    "group_id": group_id,
+                    "known_members": [
+                        {
+                            "role": "admin",
+                            "status": "active",
+                            "username": "sirosen@globusid.org",
+                        },
+                        {
+                            "role": "member",
+                            "status": "invited",
+                            "username": "bjmc@globusid.org",
+                        },
+                        {
+                            "role": "member",
+                            "status": "rejected",
+                            "username": "smartin@globusid.org",
+                        },
+                    ],
+                },
+            )
+        ),
+    )
+
+
+@pytest.fixture(autouse=True, scope="session")
 def _register_group_action_responses():
     group_id = "efdab3ca-cff1-11e4-9b86-123139260d4e"
     identity_id = "00000000-0000-0000-0000-000000000001"
