+++ tests/functional/groups/test_group_show.py
@@ -1,0 +1,17 @@
+from globus_sdk._testing import load_response_set
+
+
+def test_group_show(run_line):
+    """
+    Basic success test for globus group show
+    """
+    meta = load_response_set("cli.groups").metadata
+
+    group1_id = meta["group1_id"]
+    group1_name = meta["group1_name"]
+    group1_description = meta["group1_description"]
+
+    result = run_line(f"globus group show {group1_id}")
+
+    assert group1_name in result.output
+    assert group1_description in result.output
